+++ luigi/task.py
@@ -14,9 +14,7 @@
 
 import abc
 import logging
-import datetime
 import parameter
-import target
 import warnings
 import traceback
 import pyparsing as pp
