+++ luigi/util.py
@@ -13,8 +13,15 @@
 # the License.
 
 import warnings
+import logging
+import datetime
+from luigi import parameter
 import task
 import functools
+
+
+logger = logging.getLogger('luigi-interface')
+
 
 def common_params(task_instance, task_cls):
     """Grab all the values in task_instance that are found in task_cls"""
@@ -59,7 +66,7 @@ class inherits(object):
     def __init__(self, task_to_inherit):
         super(inherits, self).__init__()
         self.task_to_inherit = task_to_inherit
-    
+
     def __call__(self, task_that_inherits):
         this_param_names = dict(task_that_inherits.get_nonglobal_params()).keys()
         for param_name, param_obj in self.task_to_inherit.get_params():
@@ -84,7 +91,7 @@ class requires(object):
 
     def __call__(self, task_that_requires):
         task_that_requires = self.inherit_decorator(task_that_requires)
-        
+
         # Modify task_that_requres by subclassing it and adding methods
         @task_wraps(task_that_requires)
         class Wrapped(task_that_requires):
@@ -107,7 +114,7 @@ class copies(object):
     def __init__(self, task_to_copy):
         super(copies, self).__init__()
         self.requires_decorator = requires(task_to_copy)
-    
+
     def __call__(self, task_that_copies):
         task_that_copies = self.requires_decorator(task_that_copies)
 
@@ -152,7 +159,7 @@ def delegates(task_that_delegates):
         def deps(self):
             # Overrides method in base class
             return task.flatten(self.requires()) + task.flatten([t.deps() for t in task.flatten(self.subtasks())])
-        
+
         def run(self):
             for t in task.flatten(self.subtasks()):
                 t.run()
@@ -295,3 +302,46 @@ def deprecate_kwarg(old_name, new_name, kw_value):
             return function(*args, **kwargs)
         return new_function
     return real_decorator
+
+
+def previous(task):
+    """Return a previous Task of the same family.
+
+    By default checks if this task family only has one non-global parameter and if
+    it is a DateParameter, DateHourParameter or DateIntervalParameter in which case
+    it returns with the time decremented by 1 (hour, day or interval)
+    """
+    params = task.get_nonglobal_params()
+    previous_params = {}
+    previous_date_params = {}
+
+    for param_name, param_obj in params:
+        param_value = getattr(task, param_name)
+
+        if isinstance(param_obj, parameter.DateParameter):
+            previous_date_params[param_name] = param_value - datetime.timedelta(days=1)
+        elif isinstance(param_obj, parameter.DateHourParameter):
+            previous_date_params[param_name] = param_value - datetime.timedelta(hours=1)
+        elif isinstance(param_obj, parameter.DateIntervalParameter):
+            previous_date_params[param_name] = param_value.prev()
+        else:
+            previous_params[param_name] = param_value
+
+    previous_params.update(previous_date_params)
+
+    if len(previous_date_params) == 0:
+        raise NotImplementedError("No task parameter - can't determine previous task")
+    elif len(previous_date_params) > 1:
+        raise NotImplementedError("Too many date-related task parameters - can't determine previous task")
+    else:
+        return task.clone(**previous_params)
+
+
+def get_previous_completed(task, max_steps=10):
+    prev = task
+    for i in xrange(max_steps):
+        prev = previous(prev)
+        logger.debug("Checking if %s is complete" % prev.task_id)
+        if prev.complete():
+            return prev
+    return None
